requirements.txt
pymongo
numpy
pandas
python-dotenv
streamlit